requirements.txt
Django==2.2.13
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
# psycopg2>=2.7.5,<2.8.0
flake8==3.8.3